Extract RarityButtons logic into its own component
